+++ src/PluginProcessor.h
@@ -158,6 +158,7 @@ private:
   int lastAliveCount = 0;
   int stagnationCounter = 0;
   int overpopCounter = 0;
+  int reseedCooldown = 0;
   uint64_t reseedRng = 12345;
   std::atomic<uint64_t> currentSeed{12345};
   std::atomic<bool> seedChanged{false};